move 'count number of attended sessions' test to action plan repository test
